feat: add new tab for map in bottom navigator (#2945)
* feat: add new tab for map in main screen

* refactor: control show/hide map tab by remote config

* refactor: control show/hide map tab by local config

* refactor: fix position of selected position indicator

* refactor: should show selection Pin on the basis of prop

* refactor: add type to coordinates

* Update src/translations/screens/Profile.ts


* refactor: rename map tab to page
