+++ src/components/map/LocationBar.tsx
