+++ src/location-search/types.ts
@@ -1,6 +1,5 @@
 import {RootStackScreenProps} from '@atb/navigation/types';
 import {JourneySearchHistoryEntry} from '@atb/search-history/types';
-import {CompositeScreenProps} from '@react-navigation/native';
 import {StackScreenProps} from '@react-navigation/stack';
 import {
   Location,
@@ -9,6 +8,7 @@ import {
 } from '../favorites/types';
 import {RouteParams as LocationSearchRouteParams} from './LocationSearch';
 import {RouteParams as MapSelectionRouteParams} from './map-selection';
+import {CompositeScreenProps} from '@react-navigation/native';
 
 export type LocationSearchResult = {
   location: SearchLocation;
