+++ src/remote-config.ts
@@ -26,6 +26,7 @@ export type RemoteConfig = {
   enable_token_fallback: boolean;
   enable_flex_tickets: boolean;
   enable_vipps_login: boolean;
+  enable_map_page: boolean;
   favourite_departures_poll_interval: number;
   new_favourites_info_url: string;
 };
@@ -55,6 +56,7 @@ export const defaultRemoteConfig: RemoteConfig = {
   enable_token_fallback: true,
   enable_flex_tickets: false,
   enable_vipps_login: false,
+  enable_map_page: false,
   favourite_departures_poll_interval: 30000,
   new_favourites_info_url: 'https://www.atb.no',
 };
@@ -122,6 +124,10 @@ export function getConfig(): RemoteConfig {
     values['enable_vipps_login']?.asBoolean() ??
     defaultRemoteConfig.enable_vipps_login;
 
+  const enable_map_tab =
+    values['enable_map_tab']?.asBoolean() ??
+    defaultRemoteConfig.enable_map_page;
+
   const favourite_departures_poll_interval =
     values['favourite_departures_poll_interval']?.asNumber() ??
     defaultRemoteConfig.favourite_departures_poll_interval;
@@ -155,6 +161,7 @@ export function getConfig(): RemoteConfig {
     enable_token_fallback,
     enable_flex_tickets,
     enable_vipps_login,
+    enable_map_page: enable_map_tab,
     favourite_departures_poll_interval,
     new_favourites_info_url,
   };
