+++ src/translations/dictionary.ts
@@ -14,6 +14,7 @@ const dictionary = {
       `Aktiver for reisesøk. Aktiver to ganger for å nullstille søket.`,
       'Activate for search assistant. Activate twice to reset search.',
     ),
+    map: _(`Kart`, 'Map'),
     nearby: _(`Av${softhyphen}ganger`, 'Departures'),
     ticketing: _('Billetter', 'Tickets'),
     profile: _('Min profil', 'My profile'),
