+++ src/translations/screens/Profile.ts
@@ -43,6 +43,7 @@ const ProfileTexts = {
       departures: _('Prøv ny avganger-visning', 'Try the new departure view'),
       assistant: _('Prøv ny reisesøkmotor', 'Try the new travel search engine'),
       frontPage: _('Prøv ny forside', 'Try the new Front Page'),
+      map: _('Prøv ny kartside', 'Try the new Map page'),
     },
     settings: {
       heading: _('Innstillinger', 'Settings'),
